Refactor IBaseURL change package name in soegningURL og TraverserURL
