+++ src/main/java/Controller/CrawlerController/ControllerGETTilsynsrapport.java
@@ -5,12 +5,10 @@ package Controller.CrawlerController;
 import Model.Crawler.TraverserURL;
 import Model.Crawler.documentURL;
 import Model.Crawler.soegningURL;
-import Controller.Download.DownloadPDF;
 import Controller.TesxtExtraction.PDFManager;
-import Model.IBaseURL.IBaseURL;
+import Model.IBaseURL;
 import Model.DTO.Institutions.Vuggestue;
 
-import javax.ws.rs.core.Response;
 import java.util.List;
 
 public class ControllerGETTilsynsrapport {
